Converted logging file to typescript and moved to webscraper folder
