+++ wee/webscraper/src/robots/robots.service.ts
@@ -2,7 +2,7 @@ import { Injectable, Logger } from '@nestjs/common';
 import { ErrorResponse } from '../models/ServiceModels';
 import { RobotsResponse } from '../models/ServiceModels';
 // eslint-disable-next-line @nx/enforce-module-boundaries
-import logger from '../../../services/webscraperlogger';
+import logger from '../../logging/webscraperlogger';
 import fetch from 'node-fetch';
 
 const serviceName = "[RobotsService]";
