+++ wee/webscraper/src/scrape-address/scrape-address.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { RobotsResponse } from '../models/ServiceModels';
 import * as puppeteer from 'puppeteer';
-import logger from '../../../services/webscraperlogger';
+import logger from '../../logging/webscraperlogger';
 
 const serviceName = "[ScrapeAddressService]";
 @Injectable()
