+++ wee/webscraper/src/screenshot-homepage/screenshot.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { RobotsResponse, ErrorResponse } from '../models/ServiceModels';
 import * as puppeteer from 'puppeteer';
-import logger from '../../../services/webscraperlogger';
+import logger from '../../logging/webscraperlogger';
 import * as fs from 'fs/promises';
 const serviceName = "[ScreenshotService]";
 @Injectable()
